Variation that uses old OpenAI API for ChatGPT-3.5. Adds remarks.
Also remove references to identifying legal moves being a prerequisite
for playing chess. Without adversarial opponents, this may not be true
until pretty high levels, and having to argue whether having the LLM
get 10 attempts at producing a legal move, or various other ways of
assisting it means it is "playing chess" or not is rather besides
the point of this test anyway.